feat: lots of refactoring, add personal layout and analysis skeleton route
