+++ client/pages/personal/analysis.tsx
@@ -1,0 +1,9 @@
+import PersonalLayout from 'components/LayoutPersonal';
+
+export default function PersonalAnalysis() {
+  return (
+    <>
+      <PersonalLayout>The Personal Analysis page</PersonalLayout>
+    </>
+  );
+}
