+++ client/types/Transaction.ts
@@ -1,0 +1,12 @@
+import { CategoryKey } from 'data/categories';
+
+export interface Transaction {
+  id: string;
+  userId: string;
+  location: string;
+  amount: number;
+  imageUrl?: string;
+  date: number;
+  category: CategoryKey;
+  details: string;
+}
